requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-FlatPages==0.8.1
Flask-Login==0.6.2
Frozen-Flask==0.18
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
psycopg2==2.9.5
Pygments==2.14.0
PyYAML==6.0
six==1.16.0
Werkzeug==2.2.3

Pillow~=9.5.0
openpyxl~=3.1.2